🔖 Checkpoint: Edit /home/ubuntu/neo-service-layer/src/Core/NeoServiceLayer.Core/ConfidentialComputing/IEnclaveMessageBus.cs

+++ src/Core/NeoServiceLayer.Core/ConfidentialComputing/IEnclaveMessageBus.cs
@@ -842,7 +842,7 @@ namespace NeoServiceLayer.Core.ConfidentialComputing
         public string ServiceId { get; } = Guid.NewGuid().ToString();
         public string ServiceName { get; }
         public bool IsActive { get; } = true;
-        public int HandledRequestCount { get; } = 0;
+        public long HandledRequestCount { get; } = 0;
 
         public TemporaryConfidentialServiceRegistration(string serviceName)
         {
